+++ app/src/main/java/org/android/go/sopt/MainActivity.kt
@@ -4,10 +4,10 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import org.android.go.sopt.databinding.ActivityMainBinding
-import org.android.go.sopt.present.GalleryFragment
-import org.android.go.sopt.present.HomeFragment
-import org.android.go.sopt.present.MyProfileFragment
-import org.android.go.sopt.present.SearchFragment
+import org.android.go.sopt.present.menuFragment.GalleryFragment
+import org.android.go.sopt.present.menuFragment.HomeFragment
+import org.android.go.sopt.present.menuFragment.MyProfileFragment
+import org.android.go.sopt.present.menuFragment.SearchFragment
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
