+++ app/src/main/java/org/android/go/sopt/present/loginPage/IntroduceActivity.kt
@@ -1,6 +1,5 @@
-package org.android.go.sopt.myPage
+package org.android.go.sopt.present.loginPage
 
-import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
