+++ app/src/main/java/org/android/go/sopt/present/menuFragment/SearchFragment.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.present
+package org.android.go.sopt.present.menuFragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
