+++ app/src/main/java/org/android/go/sopt/remote/remoteData/datasource/LoginDataSource.kt
@@ -1,0 +1,4 @@
+package org.android.go.sopt.remote.remoteData.datasource
+
+class LoginDataSource {
+}
