+++ app/src/main/java/org/android/go/sopt/remote/service/LogInService.kt
@@ -1,7 +1,7 @@
 package org.android.go.sopt.remote.service
 
-import org.android.go.sopt.remote.model.RequestLogInDto
-import org.android.go.sopt.remote.model.ResponseLogInDto
+import org.android.go.sopt.remote.remoteData.model.RequestLogInDto
+import org.android.go.sopt.remote.remoteData.model.ResponseLogInDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
