+++ app/src/main/java/org/android/go/sopt/remote/service/UsersListService.kt
@@ -1,6 +1,6 @@
 package org.android.go.sopt.remote.service
 
-import org.android.go.sopt.remote.model.ResponseListUsersDto
+import org.android.go.sopt.remote.remoteData.model.ResponseListUsersDto
 import retrofit2.Call
 import retrofit2.http.GET
